// src/domain/ledger/shared/types.ts
export enum AccountCode {
  // TODO: rename - we should save Settlement* for accounts that get used in settlement
  Settlement_Balance = 10100,
  Deposit = 10200,
  Unrestricted = 20100,
  Clearing_Credit = 20101,
  Restricted = 20200,
  Reserved = 20300,
  Committed_Outgoing = 20400,
  Dfsp = 60100,
  Net_Debit_Cap = 60200,
  Net_Debit_Cap_Control = 60201,
  Dev_Null = 60300,
  Clearing_Setup = 60400,
  Clearing_Limit = 60500,
  Unrestricted_Lock = 60600,
  Settlement_Outgoing = 60701,
  Settlement_Incoming = 60702,

  // TODO(LD): remove me! 
  TIMEOUT = 9000,
}
// ...
export type QueryResultSuccess<T> = T extends void
  ? { type: 'SUCCESS' }
  : { type: 'SUCCESS'; result: T }

export type QueryResultFailure = {
  type: 'FAILURE',
  error: Error
}

export type QueryResultNotFound = {
  type: 'NOT_FOUND',
  error: Error
}

export type QueryResult<T> = QueryResultSuccess<T> | QueryResultFailure

// TODO: can we just combine this into QueryResult?
export type QueryResultWithNotFound<T> = QueryResultSuccess<T> | QueryResultNotFound | QueryResultFailure


/**
 * Mojaloop internal enums for things such as Transfer States, account types etc.
 * We inject these into the ledger so we don't need to keep looking them up from the db.
 */
export interface Enums {
  ledgerAccountType: Record<string, number>
  ledgerEntryType: Record<string, number>
  transferParticipantRoleType: Record<string, number>
  transferState: Record<string, number>
  participantLimitType: Record<string, number>
  settlementWindowState: Record<string, number>
  settlementDelay: Record<string, number>
  settlementDelayEnums: Record<string, number>
  settlementGranularity: Record<string, number>
  settlementGranularityEnums: Record<string, number>
  settlementInterchangeEnums: Record<string, number>
  settlementStates: Record<string, number>
  transferParticipantRoleTypes: Record<string, number>
  transferStateEnums: Record<string, number>
  transferStates: Record<string, number>
  settlementInterchange: Record<string, number>
}